+++ lab/exp-bib2bibentry.R
@@ -8,4 +8,10 @@ library(biblio)
 Bib <- read_bib(x = file.path(path.package("biblio"),
         "LuebertPliscoff.bib"))
 
-bib2bibentry(Bib[1:5, ])
+x <- bib2bibentry(Bib[1:5, ])
+
+# Back to bib
+y <- capture.output(print(x, "bibtex"))
+
+cat(paste0(y, collapse = "\n"))
+