+++ NEWS.md
@@ -11,6 +11,10 @@
 
 - Improved function `read_bib()` using regular expressions.
 
+### Bug Fixes
+
+- Function `read_bib()` was not able to read libraries with a single reference.
+
 
 # biblio 0.0.7